_requirements.txt
APScheduler
GPUtil
numpy
psutil
pandas
py-cpuinfo
PyQt5
PySide2